Organize Main branch file hierarchy, and organize RobotContainer
