+++ README.md
@@ -1,0 +1,15 @@
+# Student-Led Allegro2024
+
+This is the student-led robot code for Allegro, 1425's robot designed for the First Robotics Competition 2024 game, CRESCENDO.
+
+### Whats different about this?
+
+Most obviously, this codebase is written solely by students.
+
+That is not the only change however, for this codebase we are moving to the
+[WPILIB command-based](https://docs.wpilib.org/en/stable/docs/software/commandbased/index.html) robot framework, and experimenting with
+[AdvantageKit](https://github.com/Mechanical-Advantage/AdvantageKit) and [AdvantageScope](https://github.com/Mechanical-Advantage/AdvantageScope), Mechanical Advantage's logging framework and telemetry application.
+
+#### See Also:
+
+- [allegro2024](https://github.com/errorcodexero/allegro2024), our original code for the 2024 season.
