+++ src/main/java/frc/robot/RobotContainer.java
@@ -5,8 +5,8 @@
 package frc.robot;
 
 import com.ctre.phoenix6.Utils;
-import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
 import com.ctre.phoenix6.mechanisms.swerve.SwerveModule.DriveRequestType;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
@@ -14,19 +14,13 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
+import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.Constants.OperatorConstants;
-import frc.robot.commands.Autos;
 import frc.robot.generated.TunerConstants;
 import frc.robot.subsystems.IntakeShooter.IntakeShooterIOHardware;
 import frc.robot.subsystems.IntakeShooter.IntakeShooterSubsystem;
 import frc.robot.subsystems.Swerve.CommandSwerveDrivetrain;
-
-import com.revrobotics.CANSparkFlex;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.subsystems.oi.OIConstants;
+import frc.robot.subsystems.Swerve.util.Telemetry;
 import frc.robot.subsystems.oi.OISubsystem;
 
 
@@ -37,35 +31,37 @@ import frc.robot.subsystems.oi.OISubsystem;
  * subsystems, commands, and trigger mappings) should be declared here.
  */
 public class RobotContainer {
+
+  // Gamepad and generic initialization
+
+  private final CommandXboxController gamepad_ =
+    new CommandXboxController(OperatorConstants.kDriverControllerPort);
+
   private double MaxSpeed = TunerConstants.kSpeedAt12VoltsMps; // kSpeedAt12VoltsMps desired top speed
   private double MaxAngularRate = 1.5 * Math.PI; // 3/4 of a rotation per second max angular velocity
 
-  /* Setting up bindings for necessary control of the swerve drive platform */
-  private final CommandXboxController joystick = new CommandXboxController(0); // My joystick
-  private final CommandSwerveDrivetrain drivetrain = TunerConstants.DriveTrain; // My drivetrain
+  private final SwerveRequest.FieldCentric drive_ = new SwerveRequest.FieldCentric()
+  .withDeadband(MaxSpeed * 0.1).withRotationalDeadband(MaxAngularRate * 0.1) // Add a 10% deadband
+  .withDriveRequestType(DriveRequestType.OpenLoopVoltage); // I want field-centric driving in open loop
 
-  private final SwerveRequest.FieldCentric drive = new SwerveRequest.FieldCentric()
-      .withDeadband(MaxSpeed * 0.1).withRotationalDeadband(MaxAngularRate * 0.1) // Add a 10% deadband
-      .withDriveRequestType(DriveRequestType.OpenLoopVoltage); // I want field-centric
-                                                               // driving in open loop
-  private final SwerveRequest.SwerveDriveBrake brake = new SwerveRequest.SwerveDriveBrake();
-  private final SwerveRequest.PointWheelsAt point = new SwerveRequest.PointWheelsAt();
+  private final SwerveRequest.SwerveDriveBrake brake_ = new SwerveRequest.SwerveDriveBrake();
+  private final SwerveRequest.PointWheelsAt point_ = new SwerveRequest.PointWheelsAt();
 
-  private final Telemetry logger = new Telemetry(MaxSpeed);
+  private final Telemetry swerveTelemetry_ = new Telemetry(MaxSpeed);
 
-   // The robot's subsystems and commands are defined here...
+  // Subsystems
+
+  private final CommandSwerveDrivetrain drivetrain_ = TunerConstants.DriveTrain; // Drive train object
+
   private final IntakeShooterSubsystem intake_shooter_ = new IntakeShooterSubsystem(new IntakeShooterIOHardware());
 
   private final OISubsystem oiPanel_ = new OISubsystem(2);
-
-  // Replace with CommandPS4Controller or CommandJoystick if needed
-  private final CommandXboxController gamepad_ =
-      new CommandXboxController(OperatorConstants.kDriverControllerPort);
 
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
     // Configure the trigger bindings
     configureBindings();
+    setupDrivetrain();
   }
 
   /**
@@ -79,49 +75,29 @@ public class RobotContainer {
    */
   private void configureBindings() {
 
-    oiPanel_.setIndicator(2, false);
-    oiPanel_.setIndicator(3, true);
-    oiPanel_.setIndicator(4, false);
-
-    // Button Interactivity Testing
-    oiPanel_.climbPrepare().onTrue(Commands.runOnce(() -> {
-      oiPanel_.setIndicator(OIConstants.Indicators.climbPrepareEnabled, true);
-      oiPanel_.setIndicator(OIConstants.Indicators.climbExecuteEnabled, false);
-    }, oiPanel_));
-
-    oiPanel_.climbExecute().onTrue(Commands.runOnce(() -> {
-      oiPanel_.setIndicator(OIConstants.Indicators.climbPrepareEnabled, false);
-      oiPanel_.setIndicator(OIConstants.Indicators.climbExecuteEnabled, true);
-    }, oiPanel_));
-
-    oiPanel_.unclimb().whileTrue(Commands.runOnce(() -> {
-      oiPanel_.setIndicator(OIConstants.Indicators.unclimbEnabled, true);
-    }, oiPanel_));
-
-    oiPanel_.unclimb().whileFalse(Commands.runOnce(() -> {
-      oiPanel_.setIndicator(OIConstants.Indicators.unclimbEnabled, false);
-    }, oiPanel_));
-
-        drivetrain.setDefaultCommand( // Drivetrain will execute this command periodically
-        drivetrain.applyRequest(() -> drive.withVelocityX(-joystick.getLeftY() * MaxSpeed) // Drive forward with
-                                                                                           // negative Y (forward)
-            .withVelocityY(-joystick.getLeftX() * MaxSpeed) // Drive left with negative X (left)
-            .withRotationalRate(-joystick.getRightX() * MaxAngularRate) // Drive counterclockwise with negative X (left)
-        ));
-
-    joystick.a().whileTrue(drivetrain.applyRequest(() -> brake));
-    joystick.b().whileTrue(drivetrain
-        .applyRequest(() -> point.withModuleDirection(new Rotation2d(-joystick.getLeftY(), -joystick.getLeftX()))));
+    gamepad_.a().whileTrue(drivetrain_.applyRequest(() -> brake_));
+    gamepad_.b().whileTrue(drivetrain_
+        .applyRequest(() -> point_.withModuleDirection(new Rotation2d(-gamepad_.getLeftY(), -gamepad_.getLeftX()))));
 
     // reset the field-centric heading on left bumper press
-    joystick.leftBumper().onTrue(drivetrain.runOnce(() -> drivetrain.seedFieldRelative()));
-
-    if (Utils.isSimulation()) {
-      drivetrain.seedFieldRelative(new Pose2d(new Translation2d(), Rotation2d.fromDegrees(90)));
-    }
-    drivetrain.registerTelemetry(logger::telemeterize);
+    gamepad_.leftBumper().onTrue(drivetrain_.runOnce(() -> drivetrain_.seedFieldRelative()));
   }
 
+  private void setupDrivetrain() {
+    
+    drivetrain_.setDefaultCommand( // Drivetrain will execute this command periodically
+    drivetrain_.applyRequest(() -> drive_.withVelocityX(-gamepad_.getLeftY() * MaxSpeed) // Drive forward with
+                                                                                        // negative Y (forward)
+        .withVelocityY(-gamepad_.getLeftX() * MaxSpeed) // Drive left with negative X (left)
+        .withRotationalRate(-gamepad_.getRightX() * MaxAngularRate) // Drive counterclockwise with negative X (left)
+    ));
+
+    if (Utils.isSimulation()) {
+      drivetrain_.seedFieldRelative(new Pose2d(new Translation2d(), Rotation2d.fromDegrees(90)));
+    }
+
+    drivetrain_.registerTelemetry(swerveTelemetry_::telemeterize);
+  }
 
   /**
    * Use this to pass the autonomous command to the main {@link Robot} class.
